example_2310/requirements.txt
aioredis==1.3.1
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.2.0
audioread==2.1.9
autobahn==21.11.1
Automat==20.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.1
charset-normalizer==2.0.9
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.1
Django==3.2.10
django-allauth==0.47.0
django-appconf==1.0.5
django-braces==1.15.0
django-compressor==3.1
django-cors-headers==3.10.1
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-image-optimizer==1.0.1
django-ipware==4.0.2
django-lockdown==4.0.0
django-otp==1.1.3
django-phonenumber-field==6.0.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
django-two-factor-auth==0.1.2
django-user-sessions==1.7.1
djangochannelsrestframework==1.1.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
emoji==1.6.1
filetype==1.0.8
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
joblib==1.1.0
jsonfield==3.1.0
librosa==0.8.1
llvmlite==0.37.0
MarkupSafe==2.0.1
msgpack==1.0.3
numba==0.54.1
numpy==1.20.3
oath==1.1
oauthlib==3.1.1
packaging==21.3
phonenumbers==8.12.39
Pillow==8.4.0
pooch==1.5.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pydot-ng==2.0.0
PyJWT==1.7.1
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dotenv==1.0.1
python-resize-image==1.1.20
python-telegram-bot==13.9
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
rcssmin==1.1.0
redis==4.0.2
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rjsmin==1.2.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.1
scipy==1.7.3
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
SoundFile==0.10.3.post1
soupsieve==2.3.1
sqlparse==0.4.2
threadpoolctl==3.0.0
tinify==1.5.2
tornado==6.1
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
websocket-client==1.2.3
wrapt==1.13.3
zope.interface==5.4.0
